Added new Audio REST Api. I will update it later to add downloading music support, as it now only gives downloaded songs. Also added FfMpeg2.cs which converts the given path to StandardOutput audio.
Added a temporary web socket solution. I have to change it soon, because this solution doesn't support Nginx reverse proxy passthrough.

Added new abuse command which only I can use. Also added some moderator commands like {prefix}ban and {prefix}kick.

Made the GetPlayer method non async, because it doesn't need to be async (smh).

Changed how the FFmpeg client works. Now it has two different methods, one is to directly set the path for FFmpeg, and the other is only for Youtube videos, because of a bug that terminates the connection to the servers and skipping the last 10 seconds of any track that is over 2 minutes. The new method now caches the data in a MemoryStream and then passes it through to the StandardInput of FFmpeg, thus terminating the bug. Also added some arguments for both the FFmpeg instances.

For the HttpClient.cs class I added a custom User Agent, and added a new method called DownloadStream.

Added more information in the basic commands, because many of the player commands didn't send any return text.

For the queue class, I made it return the IPlayableItem which is removed by the RemoveFromQueue method.

I also made the Custom File Format accept file links.