Fixing some stuff :D, see the other goodies below.
Added some more spyware - I am kidding, it's only insignificant information, to know if it's safe to restart the bot.

Changes to CommandsSlash.cs: Fixed some old commands, because they were inefficient and added loop slash command with enum selection, and a move slash command

Made the GetPlayer Method take a bool to make it generate a new instance. Fixes a bug when a command is used and the bot isn't in the channel, so the method makes a new instance of the Player.cs class even when it doesn't have to. Thank god I noticed this sooner because I have a feeling it would've been a pain in the arse to fix.

In Player.cs I fixed a bug when Discord closes the audio output sink (for god knows what reason), to make it automatically reconnect when disconnected. Seeing how I mention some bugs, I should make a file containing all the bugs, this bot has had since this commit. This should be interesting.

Reworked the Vbox7 (added: SearchClient.cs and Vbox7Object.cs) functionality, because I hate Chromium, and the search taking 2 seconds to complete, and at the same time costing ~600MB ram. The new implementation takes 50ms. :D

Added Twitch Live Stream support (using YT-dlp | in the future I will shamelessly steal their code and port it to C# for maximum performance :D)

Added a new nD Audio remover because I hate these videos, as you can see by the comments in Youtube/Video.cs